AMP-10498
git-svn-id: file:///svn/branches/AMP_1_16_RELEASE@15814 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/digijava/module/aim/dbentity/AmpActivityProgram.java
@@ -5,7 +5,7 @@ import org.digijava.module.aim.util.ProgramUtil;
 public class AmpActivityProgram {
 
         private Long ampActivityProgramId;
-        private Float programPercentage;
+        private Long programPercentage;
         private AmpTheme program;
         private AmpActivity activity;
         private AmpActivityProgramSettings programSetting;
@@ -13,7 +13,7 @@ public class AmpActivityProgram {
                 return ampActivityProgramId;
         }
 
-        public Float getProgramPercentage() {
+        public Long getProgramPercentage() {
                 return programPercentage;
         }
 
@@ -33,7 +33,7 @@ public class AmpActivityProgram {
                 this.ampActivityProgramId = ampActivityProgramId;
         }
 
-        public void setProgramPercentage(Float programPercentage) {
+        public void setProgramPercentage(Long programPercentage) {
                 this.programPercentage = programPercentage;
         }
